+++ Converter-Microservice/DataAccess/Repository/QueueRepository.cs
@@ -175,7 +175,8 @@ namespace DataAccess.Repository
             if (converterResult.Wait(60000))
             {
                 e.Model.BasicAck(deliveryTag: ea.DeliveryTag, multiple: false);
-                File.Delete(objModel.FileFullPath);
+                if (File.Exists(objModel.FileFullPath))
+                    File.Delete(objModel.FileFullPath);
             }
 
             QueueLog queueLog = new QueueLog()
